Pre: Split the ever growing webcontents sagas into two files: commands and events
